Making sending from a separate thread work when using Woo as a backend server.
In current websocket-driver there is a problem - when you start a separate thread on the server and trying to 'send' message to the client, it is delivered with delay after only 60 seconds. The issue exists only when using Woo server. On Hunchentoot, message is delivered immediately.

The problem is the libev's event loop. This commit is a hack around and should be improved to work with all backend servers.