Replace ironclad dependency with babel and sha1
Ironclad is a meaty dependency. This system takes the longest time on my
computer to compile my web application, and I've read comments online
that it adds ~19MB RAM usage in a lisp image. `websocket-driver` is
currently using ironclad for two functions:

- Convert ASCII strings to octets
- Create SHA1 digests

Babel, a transient dependency, already provides this. Ironclad's only
filling need, then, is SHA1 hashing.

This change switches ironclad with the single-purpose library `sha1` to
minimize websocket-driver's footprint. It also promotes babel from a
transient to a direct dependency, in the process.

Signed-off-by: Samuel Hunter <samuel@shunter.xyz>