Emit a socket closed event in a non-async socket loop
Normally when using an async web server, websocket-driver handles broken
connections by attempting to write to a stream, and if that fails, it
closes the socket and emits the close event. In the case of a non-async
webserver like Hunchentoot, a loop form checks the stream's status
manually, and then closes the socket once the loop exits. But in this
case, it doesn't emit the close event. This PR adds the event to this
case as well.